fix: 댓글 및 리뷰페이지관련 오류해결

+++ src/main/java/org/team14/webty/recommend/service/RecommendService.java
@@ -66,7 +66,7 @@ public class RecommendService {
 
 	public Map<String, Boolean> isRecommended(WebtyUserDetails webtyUserDetails, Long reviewId) {
 		WebtyUser webtyUser = webtyUserDetails.getWebtyUser();
-		Map<String, Integer> rawResult = recommendRepository.findRecommendStatusByUserAndReview(webtyUser.getUserId(),
+		Map<String, Long> rawResult = recommendRepository.findRecommendStatusByUserAndReview(webtyUser.getUserId(),
 			reviewId);
 		return rawResult.entrySet().stream()
 			.collect(Collectors.toMap(
